% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interestRateModels.R
\name{ann.ev}
\alias{ann.ev}
\title{Expected value of annuity}
\usage{
ann.ev(n, irm)
}
\arguments{
\item{n}{The length of the annuity}

\item{irm}{The interest rate model to use}
}
\value{
Present value of annuity
}
\description{
Calculates the expected value of an n-period annuity.
See the \link{iratemodel} class for details and examples.
}

