% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interestRateModels.R
\name{ann.var}
\alias{ann.var}
\title{Variance of annuity}
\usage{
ann.var(n, irm)
}
\arguments{
\item{n}{The length of the annuity}

\item{irm}{The interest rate model to use}
}
\value{
Variance of annuity
}
\description{
Calculates the variance of an n-period annuity
}

