% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interestRateModels.R
\name{ar1}
\alias{ar1}
\alias{delta.cov.ar1}
\alias{delta.ev.ar1}
\alias{delta.var.ar1}
\alias{y.cov.ar1}
\alias{y.ev.ar1}
\alias{y.var.ar1}
\title{The AR(1) Process}
\usage{
\method{delta.ev}{ar1}(t, irm)

\method{delta.cov}{ar1}(s, t, irm)

\method{delta.var}{ar1}(t, irm)

\method{y.ev}{ar1}(t, irm)

\method{y.var}{ar1}(t, irm)

\method{y.cov}{ar1}(s, t, irm)
}
\description{
A class used to model an AR(1) process. See the \link{iratemodel}
class for details and examples.
}
\references{
Parker, Gary. An application of stochastic interest rate models in
  life assurance. Diss. Heriot-Watt University, 1992.
}

