% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interestRateModels.R
\name{ar1process}
\alias{ar1process}
\alias{delta.cov.ar1}
\alias{delta.ev.ar1}
\alias{delta.var.ar1}
\alias{y.cov.ar1}
\alias{y.ev.ar1}
\alias{y.var.ar1}
\title{The AR(1) process Process}
\usage{
\method{delta.ev}{ar1}(t, irm)

\method{delta.cov}{ar1}(s, t, irm)

\method{delta.var}{ar1}(t, irm)

\method{y.ev}{ar1}(t, irm)

\method{y.var}{ar1}(t, irm)

\method{y.cov}{ar1}(s, t, irm)
}
\value{
\code{y.ev} gives the expected value, \code{y.var}
gives the variance and \code{y.cov} gives the covariance.
}
\description{
Expected value, variance and covariance for the Ornstein-Uhlenbeck 
process with current force of interest equal to \code{delta0}, ultimate force of
interest equal to \code{delta}, friction parameter equal to \code{alpha},
and diffusion coefficient equal to \code{sigma}.
}
\details{
The Geometric Brownian Motion process is defined by

\deqn{delta_t = \delta_{0} + \delta_s - \delta) dt + \sigma dW_t}

where \eqn{W_t} is a standard Wiener process with parameter \eqn{\sigma = 1}
and \eqn{\alpha}, \eqn{\sigma} and \eqn{\delta} are constants with
\eqn{\alpha \ge 0} and \eqn{\sigma \ge 0}.

Define \eqn{y(t)} as

\deqn{y(t) = \int_{0}^{t} \delta_s ds}
}
\references{
Parker, Gary. An application of stochastic interest rate models in
  life assurance. Diss. Heriot-Watt University, 1992.
}

