% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interestRateModels.R
\name{arma}
\alias{arma}
\alias{delta.cov.arma}
\alias{delta.ev.arma}
\alias{delta.var.arma}
\alias{y.cov.arma}
\alias{y.ev.arma}
\alias{y.var.arma}
\title{The ARMA(2,1) Process}
\usage{
\method{delta.ev}{arma}(t, irm)

\method{delta.cov}{arma}(s, t, irm)

\method{delta.var}{arma}(t, irm)

\method{y.ev}{arma}(t, irm)

\method{y.var}{arma}(t, irm)

\method{y.cov}{arma}(s, t, irm)
}
\description{
A class used to model the ARMA(2,1) process. See the \link{iratemodel}
class for details and examples.
}
\references{
Parker, Gary. An application of stochastic interest rate models in
  life assurance. Diss. Heriot-Watt University, 1992.
}

