% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insuranceModels.R
\docType{data}
\name{CdnPop2016}
\alias{CdnPop2016}
\title{Canadian Population Table for 2016}
\format{Male and Female population counts and proportions for 
different age groups.}
\source{
Statistics Canada
}
\usage{
CdnPop2016
}
\description{
Canadian Population for 2016 from Statistics Canada census.
}

