% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interestRateModels.R
\name{determ}
\alias{delta.cov.determ}
\alias{delta.ev.determ}
\alias{delta.var.determ}
\alias{determ}
\alias{y.cov.determ}
\alias{y.ev.determ}
\alias{y.var.determ}
\title{Deterministic interest rate}
\usage{
\method{delta.ev}{determ}(t, irm)

\method{delta.cov}{determ}(s, t, irm)

\method{delta.var}{determ}(t, irm)

\method{y.ev}{determ}(t, irm)

\method{y.var}{determ}(t, irm)

\method{y.cov}{determ}(s, t, irm)
}
\description{
A class used to model the deterministic interest rate. 
See the \link{iratemodel} class for details and examples.
}
\references{
Parker, Gary. An application of stochastic interest rate models in
  life assurance. Diss. Heriot-Watt University, 1992.
}

