% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insuranceModels.R
\name{endowport}
\alias{endowport}
\alias{z.moment.iport.endowport}
\title{Endowment insurance portfolio (identical policies)}
\usage{
\method{z.moment}{iport.endowport}(moment, ins, mort, irm)
}
\description{
A portfolio of \code{c} identical endowment policies. See the 
\link{insurance} class for details and examples on how to use this class.
}
\references{
Parker, Gary. An application of stochastic interest rate models in
  life assurance. Diss. Heriot-Watt University, 1992.
}

