% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insuranceModels.R
\name{endowsingle}
\alias{endowsingle}
\alias{z.ev.three.isingle.endowsingle}
\alias{z.ev.two.isingle.endowsingle}
\alias{z.ev.twoone.isingle.endowsingle}
\alias{z.insrisk.isingle.endowsingle}
\alias{z.invrisk.isingle.endowsingle}
\alias{z.moment.isingle.endowsingle}
\title{Single endowment insurance product}
\usage{
\method{z.moment}{isingle.endowsingle}(moment, ins, mort, irm)

z.ev.two.isingle.endowsingle(ins, mort, irm)

z.ev.three.isingle.endowsingle(ins, mort, irm)

z.ev.twoone.isingle.endowsingle(ins, mort, irm)

\method{z.insrisk}{isingle.endowsingle}(ins, mort, irm)

\method{z.invrisk}{isingle.endowsingle}(ins, mort, irm)
}
\description{
An endowment insurance pays a benefit of \code{d} if the policyholder
dies within the term of the contract and \code{e} if the policyholder
survives the term of the contract.

In addition to the functions described in the \link{insurance} class, other
functions are available for the endowsingle class. 

In particular, 
\code{z.ev.two.isingle.endowsingle} can be used to calculate \eqn{E[Z_{1}Z_{2}]},
\code{z.ev.three.isingle.endowsingle} can be used to calculate
\eqn{E[Z_{1}Z_{2}Z_{3}]} and \code{z.ev.twoone.isingle.endowsingle} 
can be used to calculate \eqn{E[Z_{1}^2Z_{2}]} in Parker (1992).
}
\details{
The \code{z.insrisk} and \code{z.invrisk} functions for the
\code{endowsingle} class are implemented such that we are conditioning
on \eqn{K} not \eqn{y}. See Parker (1997) for details.
}
\examples{
oumodel = iratemodel(list(delta0 = 0.1, delta = 0.06, 
alpha = 0.1, sigma = 0.01), "ou")
mort = mortassumptions(list(x = 40, table = "MaleMort91"))
endow = insurance(list(n = 10, d = 1, e = 1), "isingle", "endow")

z.ev(endow,mort,oumodel) # first moment
z.ev.two.isingle.endowsingle(endow, mort, oumodel)
z.ev.twoone.isingle.endowsingle(endow, mort, oumodel)
z.ev.three.isingle.endowsingle(endow, mort, oumodel)

}
\references{
Parker, Gary. An application of stochastic interest rate models in
  life assurance. Diss. Heriot-Watt University, 1992.
  
  Parker, G. (1997). Stochastic analysis of the interaction between
  investment and insurance risks. North American actuarial journal, 1(2), 55–71.
}

