% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insuranceModels.R
\docType{data}
\name{FemaleMort82Reduced}
\alias{FemaleMort82Reduced}
\title{CA Male Mortality rates from 1980-1982 * 0.75}
\format{a \code{x} column for the age from 0 to 102 inclusive
and a \code{qx} column for the mortality rate.}
\usage{
FemaleMort82Reduced
}
\description{
This data set contains the mortality rates used in Parker.
}
\references{
Parker, G. (1997). Stochastic analysis of the interaction between
investment and insurance risks. North American actuarial journal, 1(2), 55–71.
}

