% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insuranceModels.R
\name{igroup}
\alias{igroup}
\alias{z.insrisk}
\alias{z.invrisk}
\alias{z.moment.igroup}
\title{Group of Insurance Portfolios}
\usage{
\method{z.moment}{igroup}(moment, ins, mort, irm)

z.insrisk(ins, mort, irm)

z.invrisk(ins, mort, irm)
}
\description{
An endowment insurance pays a benefit of \code{d} if the policyholder
dies within the term of the contract and a benefit of \code{e} if the policyholder
survives the term of the contract.
}
\references{
Parker, Gary. An application of stochastic interest rate models in
  life assurance. Diss. Heriot-Watt University, 1992.
}

