% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insuranceModels.R
\name{iport}
\alias{iport}
\alias{z.insrisk.iport}
\alias{z.invrisk.iport}
\alias{z.moment.iport}
\alias{z.pdf.iport}
\title{Insurance portfolio (identical policies)}
\usage{
z.moment.iport(moment, ins, mort, irm)

\method{z.insrisk}{iport}(ins, mort, irm)

\method{z.invrisk}{iport}(ins, mort, irm)

\method{z.pdf}{iport}(z, ins, mort, irm)
}
\description{
A portfolio of \code{c} identical term or endowment policies.
The \link{termport} and \link{endowport} classes implement the functions for the
present value of benefit random variable. See the \link{insurance} class
for details and examples on how to use this class.
}
\references{
Parker, Gary. An application of stochastic interest rate models in
  life assurance. Diss. Heriot-Watt University, 1992.
}

