% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insuranceModels.R
\docType{data}
\name{MaleMort91}
\alias{MaleMort91}
\title{CA Male Mortality rates from 1990-1992}
\format{a \code{x} column for the age from 0 to 100 inclusive
and a \code{qx} column for the mortality rate.}
\source{
Statistics Canada
}
\usage{
MaleMort91
}
\description{
This data set contains the mortality rates from Statistics Canada.
}

