% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insuranceModels.R
\name{mortassumptions}
\alias{kdeferredqx}
\alias{kdeferredqx.mortassumptions}
\alias{kpx}
\alias{kpx.mortassumptions}
\alias{mortassumptions}
\title{Mortality Assumptions}
\usage{
mortassumptions(params)

kpx(k, mort)

kdeferredqx(k, mort)

\method{kpx}{mortassumptions}(k, mort)

\method{kdeferredqx}{mortassumptions}(k, mort)
}
\description{
An endowment insurance pays a benefit of \code{d} if the policyholder
dies within the term of the contract and a benefit of \code{e} if the policyholder
survives the term of the contract.
}
\references{
Parker, Gary. An application of stochastic interest rate models in
  life assurance. Diss. Heriot-Watt University, 1992.
}

