% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interestRateModels.R
\name{ou}
\alias{delta.cov.ou}
\alias{delta.ev.ou}
\alias{delta.var.ou}
\alias{ou}
\alias{y.cov.ou}
\alias{y.ev.ou}
\alias{y.var.ou}
\title{The Ornstein-Uhlenbeck Process}
\usage{
\method{delta.ev}{ou}(t, irm)

\method{delta.cov}{ou}(s, t, irm)

\method{delta.var}{ou}(t, irm)

\method{y.ev}{ou}(t, irm)

\method{y.var}{ou}(t, irm)

\method{y.cov}{ou}(s, t, irm)
}
\description{
A class used to model the Ornstein-Uhlenbeck process. See the \link{iratemodel}
class for details and examples.
}
\references{
Parker, Gary. An application of stochastic interest rate models in
  life assurance. Diss. Heriot-Watt University, 1992.
}

