% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interestRateModels.R
\name{ouprocess}
\alias{delta.ev.ou}
\alias{delta.var.ou}
\alias{ouprocess}
\alias{y.cov.ou}
\alias{y.ev.ou}
\alias{y.var.ou}
\title{The Ornstein-Uhlenbeck Process}
\usage{
\method{delta.ev}{ou}(t, irm)

\method{delta.var}{ou}(t, irm)

\method{y.ev}{ou}(t, irm)

\method{y.var}{ou}(t, irm)

\method{y.cov}{ou}(s, t, irm)
}
\value{
\code{y.ev} gives the expected value, \code{y.var}
gives the variance and \code{y.cov} gives the covariance.
}
\description{
Expected value, variance and covariance for the Ornstein-Uhlenbeck 
process with current force of interest equal to \code{delta0}, ultimate force of
interest equal to \code{delta}, friction parameter equal to \code{alpha},
and diffusion coefficient equal to \code{sigma}.
}
\details{
The Ornstein-Uhlenbeck process is defined by

\deqn{d\delta_t = -\alpha (\delta_t - \delta) dt + \sigma dW_t}

where \eqn{W_t} is a standard Wiener process with parameter \eqn{\sigma = 1}
and \eqn{\alpha}, \eqn{\sigma} and \eqn{\delta} are constants with
\eqn{\alpha \ge 0} and \eqn{\sigma \ge 0}.

Define \eqn{y(t)} as

\deqn{y(t) = \int_{0}^{t} \delta_s ds}
}
\references{
Parker, Gary. An application of stochastic interest rate models in
  life assurance. Diss. Heriot-Watt University, 1992.
}

