% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interestRateModels.R
\docType{data}
\name{returns91}
\alias{returns91}
\title{Investment Returns from 1991 to 2017}
\format{A column for month and column for return (annualized). Rates are
continuously compounded.}
\usage{
returns91
}
\description{
Company returns observed on monthly basis from 1991 to 2017
}

