% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interestRateModels.R
\name{second}
\alias{delta.cov.second}
\alias{delta.ev.second}
\alias{delta.var.second}
\alias{second}
\alias{y.ev.second}
\alias{y.var.second}
\title{Second Order Stochastic Differential Equation}
\usage{
\method{delta.ev}{second}(t, irm)

\method{delta.cov}{second}(s, t, irm)

\method{delta.var}{second}(t, irm)

\method{y.ev}{second}(t, irm)

\method{y.var}{second}(t, irm)
}
\description{
A class used to model the Second Order Stochastic Differential Equation
for the force of interest. See the \link{iratemodel} class for details and examples.
}
\references{
Parker, Gary. An application of stochastic interest rate models in
  life assurance. Diss. Heriot-Watt University, 1992.
}

