% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insuranceModels.R
\name{termport}
\alias{termport}
\alias{z.moment.iport.termport}
\title{Term insurance portfolio (identical policies)}
\usage{
\method{z.moment}{iport.termport}(moment, ins, mort, irm)
}
\description{
A portfolio of \code{c} identical term policies. See the 
\link{insurance} class for details and examples on how to use this class.
}
\references{
Parker, Gary. An application of stochastic interest rate models in
  life assurance. Diss. Heriot-Watt University, 1992.
}

