% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interestRateModels.R
\name{wiener}
\alias{delta.cov.wiener}
\alias{delta.ev.wiener}
\alias{delta.var.wiener}
\alias{wiener}
\alias{y.cov.wiener}
\alias{y.ev.wiener}
\alias{y.var.wiener}
\title{The Wiener Process}
\usage{
\method{delta.ev}{wiener}(t, irm)

\method{delta.cov}{wiener}(s, t, irm)

\method{delta.var}{wiener}(t, irm)

\method{y.ev}{wiener}(t, irm)

\method{y.cov}{wiener}(s, t, irm)

\method{y.var}{wiener}(t, irm)
}
\description{
A class used to model the Wiener process. See the \link{iratemodel}
class for details and examples.
}
\references{
Parker, Gary. An application of stochastic interest rate models in
  life assurance. Diss. Heriot-Watt University, 1992.
}

