% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insuranceModels.R
\docType{data}
\name{CdnPop2016}
\alias{CdnPop2016}
\title{Canadian Population Table for 2016}
\format{female and male population counts and proportions for 
different age groups.}
\source{
Appendix B in Parker.
}
\usage{
CdnPop2016
}
\description{
Canadian Population for 2016 from Statistics Canada census.
}
\references{
Parker, Gary. An application of stochastic interest rate models in
  life assurance. Diss. Heriot-Watt University, 1992.
}

