% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insuranceModels.R
\name{endowport}
\alias{endowport}
\alias{z.moment.iport.endowport}
\title{Endowment insurance portfolio}
\usage{
\method{z.moment}{iport.endowport}(moment, ins, mort, irm)
}
\description{
An endowment insurance pays a benefit of \code{d} if the policyholder
dies within the term of the contract and a benefit of \code{e} if the policyholder
survives the term of the contract.
}
\references{
Parker, Gary. An application of stochastic interest rate models in
  life assurance. Diss. Heriot-Watt University, 1992.
}

