% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insuranceModels.R
\name{igroup}
\alias{cashflow.cov}
\alias{cashflow.ev}
\alias{igroup}
\alias{z.ev.two.igroup}
\alias{z.insrisk.igroup}
\alias{z.invrisk.igroup}
\alias{z.moment.igroup}
\title{Group of Insurance Portfolios}
\usage{
z.ev.two.igroup(ind1, ind2, ins, mort, irm)

\method{z.moment}{igroup}(moment, ins, mort, irm)

cashflow.ev(r, ins, mort, irm)

cashflow.cov(s, r, ins, mort, irm)

\method{z.insrisk}{igroup}(ins, mort, irm)

\method{z.invrisk}{igroup}(ins, mort, irm)
}
\description{
A class for calculating the moments, insurance risk and
and investment risk for a group of endowment and term insurance portfolios
using the formulas described in Parker (1997).

In addition to the functions described in the \link{insurance} class some
other functions are available for the \code{igroup} class.

In particular,
\code{z.ev.two.igroup} calculates \eqn{E[Z_{1,i}, Z_{i,r}]} for
\code{i} = \code{ind1} and \code{r} = \code{ind2}, \code{cashflow.ev}
calculates the expected cashflow at time \code{r} and \code{cashflow.cov}
calculates the covariance between the cashflows at time \code{r} and time
\code{s} as describd in Parker (1997).

For details and examples on how to use the \code{igroup} class
see the \link{insurance} class.
}
\examples{
oumodel = iratemodel(list(delta0 = 0.1, delta = 0.06, 
alpha = 0.1, sigma = 0.01), "ou")
mort = mortassumptions(list(x = 40, table = "MaleMort91"))
mort2 = mortassumptions(list(x = 50, table = "FemaleMort91"))

termins = insurance(list(n = 10, d = 1), "isingle", "term")
endowins = insurance(list(n = 10, e = 1, d = 1), "isingle", "endow")

termport = insurance(list(single = termins, c = 1000), "iport", "term")
endowport = insurance(list(single = endowins, c = 1000), "iport", "endow")

groupins = insurance(list(termport, endowport), 
"igroup") # 1000 term contracts, 1000 endow contracts
groupmort = list(mort, mort2) # term contracts are age 40, endow contracts are age 50
z.moment(1, groupins, groupmort, oumodel) / groupins$c # average cost per policy
z.ev.two.igroup(1, 2, groupins, groupmort, oumodel)
cashflow.ev(5, groupins, groupmort, oumodel)
cashflow.cov(3, 5, groupins, groupmort, oumodel)

}
\references{
Parker, G. (1997). Stochastic analysis of the interaction between
  investment and insurance risks. North American actuarial journal, 1(2), 55–71.
}

