% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insuranceModels.R
\name{insurance}
\alias{insurance}
\alias{z.ev}
\alias{z.insrisk}
\alias{z.invrisk}
\alias{z.moment}
\alias{z.pdf}
\alias{z.sd}
\alias{z.sk}
\title{Insurance product}
\usage{
insurance(params, class, subclass = NULL)

z.moment(moment, ins, mort, irm)

z.ev(ins, mort, irm)

z.sd(ins, mort, irm)

z.sk(ins, mort, irm)

z.insrisk(ins, mort, irm)

z.invrisk(ins, mort, irm)

z.pdf(z, ins, mort, irm)
}
\description{
An endowment insurance pays a benefit of \code{d} if the policyholder
dies within the term of the contract and a benefit of \code{e} if the policyholder
survives the term of the contract.
}
\references{
Parker, Gary. An application of stochastic interest rate models in
  life assurance. Diss. Heriot-Watt University, 1992.
}

