% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interestRateModels.R
\name{iratemodel}
\alias{delta.ev}
\alias{delta.var}
\alias{iratemodel}
\alias{pv.cov}
\alias{pv.ev}
\alias{pv.moment}
\alias{pv.var}
\alias{y.cov}
\alias{y.ev}
\alias{y.var}
\title{Interest Rate Model}
\usage{
iratemodel(params, class)

delta.ev(t, irm)

delta.var(t, irm)

y.ev(t, irm)

y.var(t, irm)

y.cov(s, t, irm)

pv.moment(t, moment, irm)

pv.ev(t, irm)

pv.var(t, irm)

pv.cov(s, t, irm)
}
\arguments{
\item{params}{A list with the parameter values to use}

\item{class}{Either "ou" or "gbm"}
}
\description{
The following \code{iratemodel} classes are available:

- \code{ou}: Ornstein-Uhlenbeck process

- \code{gbm}: Geometric Brownian motion

Each class can be used to model the force of interest, \eqn{\delta_{s}}.
}
\details{
For information on how to use the \code{iratemodel} function see
the examples.

Define:

\deqn{y(t) = int_{0}^{t} \delta_{s}}
\deqn{pv(t) = exp{-y(t)}}

The functions \code{y.ev}, \code{y.var} and \code{y.cov} are available for each class.
}
\references{
Parker, Gary. An application of stochastic interest rate models in
  life assurance. Diss. Heriot-Watt University, 1992.
}

