% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insuranceModels.R
\name{isingle}
\alias{isingle}
\alias{z.insrisk.isingle}
\alias{z.invrisk.isingle}
\alias{z.moment.isingle}
\alias{z.pdf.isingle}
\alias{z.pdf.isingle.endowsingle}
\title{Single insurance product}
\usage{
z.moment.isingle(moment, ins, mort, irm)

z.insrisk.isingle(ins, mort, irm)

z.invrisk.isingle(ins, mort, irm)

z.pdf.isingle(z, ins, mort, irm)

\method{z.pdf}{isingle.endowsingle}(z, ins, mort, irm)
}
\description{
An endowment insurance pays a benefit of \code{d} if the policyholder
dies within the term of the contract and a benefit of \code{e} if the policyholder
survives the term of the contract.
}
\references{
Parker, Gary. An application of stochastic interest rate models in
  life assurance. Diss. Heriot-Watt University, 1992.
}

