% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insuranceModels.R
\name{mortassumptions}
\alias{kdeferredqx}
\alias{kdeferredqx.mortassumptions}
\alias{kpx}
\alias{kpx.mortassumptions}
\alias{mortassumptions}
\title{Mortality Assumptions}
\usage{
mortassumptions(params)

kpx(k, mort)

kdeferredqx(k, mort)

\method{kpx}{mortassumptions}(k, mort)

\method{kdeferredqx}{mortassumptions}(k, mort)
}
\description{
A class which describes the mortality for a policyholder.
The \code{params} argument should be a list with \code{x} for the age and
\code{table} for the table. See the examples below.
}
\details{
The \code{kpx} function can be used to calculate survival
probabilities and the \code{kdeferredqx} function can be used to calculate
probability of death in a given year. See the examples below.
}
\examples{
malemort = mortassumptions(list(x = 50, table = "MaleMort91"))
femalemort = mortassumptions(list(x = 30, table = "FemaleMort91"))

# calculate probabilities
kpx(5, malemort)
kdeferredqx(2, femalemort)
}

