% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insuranceModels.R
\name{termsingle}
\alias{termsingle}
\alias{z.ev.three.isingle.termsingle}
\alias{z.ev.two.isingle.termsingle}
\alias{z.ev.twoone.isingle.termsingle}
\alias{z.insrisk.isingle.termsingle}
\alias{z.invrisk.isingle.termsingle}
\alias{z.moment.isingle.termsingle}
\title{Single term insurance product}
\usage{
\method{z.moment}{isingle.termsingle}(moment, ins, mort, irm)

z.ev.two.isingle.termsingle(ins, mort, irm)

z.ev.three.isingle.termsingle(ins, mort, irm)

z.ev.twoone.isingle.termsingle(ins, mort, irm)

\method{z.insrisk}{isingle.termsingle}(ins, mort, irm)

\method{z.invrisk}{isingle.termsingle}(ins, mort, irm)
}
\description{
An endowment insurance pays a benefit of \code{d} if the policyholder
dies within the term of the contract and a benefit of \code{e} if the policyholder
survives the term of the contract.
}
\references{
Parker, Gary. An application of stochastic interest rate models in
  life assurance. Diss. Heriot-Watt University, 1992.
}

